# perf: if argument order (mops)